blobserver/b2: update b2 library to refresh tokens and tune blobserver
Previously the b2 library would not refresh authorization token,
resulting in errors after 24 hours of uptime.

Also, fix RemoveBlobs not to return an error on not found names. This
was working correctly with a previous iteration of the b2 library, and
was not fixed when the API changed.

github.com/FiloSottile/b2: d05beb60354dacaa61460156f5c498bc25aecbcf

Change-Id: I8d735abd64616b82fd95807e15ffdd7d450a4ed6