make.go: always populate gopherjsGoroot
The change in I7c80976d46ce8d69e394d226c685c8554176ea0f wasn't working
for me.  make.go might be a little smarter by requiring the exact Go
version that it knows the pinned gopherjs needs, instead of just
checking it is not devel, but in any case it seems sensible to always
respect the env var.

Successfully built like this after this patch:

    CAMLI_GOPHERJS_GOROOT=$(go env GOROOT) go1.8rc3 run make.go

Change-Id: I58f398531cd65847b5bf82f9b9d0aa26bbe9b90d