website: fix nil based panic
The parent's fileInfo (outer block) was replaced with the child's
fileInfo (in the for loop) even when os.Lstat failed, i.e. when the latter
fileInfo was actually nil. The next fi.IsDir() (outer block) call would
then panic.

This change only updates the fileInfo when os.Lstat did not fail (i.e.
when a child was actually found).

Change-Id: Id508981354e661b6e8d6f2a0ef6818fbd6b7393e