pkg/index: simplify out of order indexing
There's a race that takes place at the end of the reindexing process.

In func (x *Index) Reindex(), we wait for all the reindexing goroutines
to be done, with wg.Wait. However, any (or all) of those goroutines
could have triggered (they call indexBlob, which calls
blobserver.Receive, which calls noteBlobIndexed, which can send on
tickleOoo) an asynchronous out of order reindexing which is NOT waited
on.

The race can be trivially demonstrated by changing:

  WaitTickle:
  	for range ix.tickleOoo {
+  		time.Sleep(5*time.Second)

in receive.go, and running:
  go test ./pkg/index/ -run TestReindex_*

This CL rewrites the out of order indexing implementation to make it
simpler, and in the process, fixes the above bug.

Fixes #756

Change-Id: If79fb1ad8869cefce4a095ef2becdba333732bce