devcam/server: update default config with prefix
This update the default configuration such that `devcam server` and
`devcam server -hello` work again. Changes in whether prefix and listen
were required on the application left both of those targets broken.

The default server target was only broken briefly, when the scanning
cabinet was introduced.

The `server -hello` command has been broken for months.

This change is a temporary one. A more proper fix is present in
https://camlistore-review.googlesource.com/#/c/9206/.

Once that change is merged, this commit can be reverted.

Change-Id: If692760da265502fece3b981bdd34d4d5bea5690