pkg/index: ignore unset msdos time when possible
If a zip archive is created without specifying the modtimes of the
files, they'll end up with a default modtime set to the MSDOS epoch
(1980-01-01 modulo some timezone and silly details), which is a common
enough occurrence.

Even when the index has a better information, such as the EXIF time,
when clients of the index (the web UI, through the search package) sort
by creation time, they use the oldest indexed time available, which is
unfortunate in that case.

Therefore, this CL makes the indexer ignore the oldest time found, if it
is before the MSDOS epoch, and if we have another time available, when
receiving a file.

Also fixed the use of hardcoded value of keyFileTimes.name, to help with
reading/searching code.

Change-Id: I9c2c39b319fdf6cd5214cab8928dd025451077ac