server/camlistored/ui: enable multi-items sharing
Sharing through the web UI was limited to a selection of one (permanode)
file or directory.

This change enables sharing of all the selected items, as long as they
are (permanodes of) files or directories, by creating a new directory
blob containing the selected items. The new directory is the item
getting shared. For clarity, and hopefully helping with search later,
the new directory is named "shared-YYYYMMDDhhmmss".

Change-Id: Iabd2266c9af85f3e8fbfa56ce98f653c0ce0f1ee