camlistored/ui: connect temp form to DOM before submitting
The download action of the web UI uses an ephemeral form, that we do not
connect to the DOM, behind the scenes. However, according to
https://html.spec.whatwg.org/multipage/forms.html#form-submission-algorithm,
step 2., a form must be connected to the DOM in order for it to be
submitted.

This change connects the form right before submitting, and removes it
from the DOM right after submission.

Fixes #921

Change-Id: I4d993c633c60a369ce4260f84731a64e8b955c76