server/camlistored: listen for SIGTERM
Systemd by default sends SIGTERM to child processes when shutting down
or restarting a service, which is not handled by camlistored and results
in a hard kill.

This change makes camlistored listen for SIGTERM next to SIGINT to allow
graceful shutdowns. The log message for SIGHUP is adjusted for
consistency.

Change-Id: Ia9be84b0b2284343e1776042fdfc99363110bfb1