clients/android: migrate to gradle
Android applications nowadays are supposed to be developped with Android
Studio, which uses gradle as the build system, with the android plugin.

This change modifies the app setup and file tree layout so it can be
built with gradle (in docker by default) instead of with ant.

The initial motivation for this change is to make changes to the app in
Go, using 'gomobile bind' to call the Go code from java. As the canonical
example at golang.org/x/mobile/example/bind/hello uses gradle, the
first step was to migrate the app to gradle.

Even if we don't end up using gomobile, this change is interesting as it
makes it easier for android apps developpers to contribute to the app if
it uses the recommended "modern" way.

Change-Id: I8386012db640324b62e18ffc128b68ba971fa953