vendor: upgrade gopherjs to dc374d3
rev dc374d32704510cb387457180ca9d5193978b555

Upgrade required for move to myitcv.io/react (because it is tested
against this version) but also generally a good idea to stay up to date
with the latest version of GopherJS... it's not totally stable (yet)

Gopherjs now depends on the commonly used github.com/kisielk/gotool, a
library of some of the utility functions provided by (but not exported)
by cmd/go. Gopherjs uses this library to expand package specifications
provided to the compiler, i.e. ./..., in a way consistent with cmd/go.

github.com/kisielk/gotool is pinned at rev
0de1eaf82fa3f583ce21fde859f1e7e0c5e9b220

Within Gopherjs we are using fsnotify workaround described in
https://camlistore-review.googlesource.com/c/9409

Fixes #930

Change-Id: I449fa0c6fccf4abed69304cc17eeb75c83a0235b