server/camlistored/ui: add map aspect
To display permanodes with a location on a map graphical interface (like
Google Maps, Open Street Map, etc).

Screenshots:
https://storage.googleapis.com/camlistore-screenshots/Screenshot_20170620-230922.png
https://storage.googleapis.com/camlistore-screenshots/Screenshot_20170620-231221.png
https://storage.googleapis.com/camlistore-screenshots/Screenshot_20170620-235006.png

Lots more features possible, but I think this is a good point to add it
as an mvp. TODOs in the code.

Using leaflet.js version 1.0.3, and tiles from mapbox.com (Open Street
Map).

Change-Id: Ib31610c122de633c5268174928c22c1cc481f388