clients/android: avoid wifi manager memory leak
When building the uploader application with a newer version of the build
tools, the build fails in stage app:lintVitalRelease with a lint error of
type WifiMangerLeak. The wifi manager instance should always be obtained
from an application context in order to avoid a memory leak on Android
versions before N. See:
https://developer.android.com/reference/android/net/wifi/WifiManager.html

This commit implements the recommendation to obtain the wifi manager
instance through the application context.

Change-Id: I59ff0dfb381b82d9b4f201b646f11b34b3de3cff