pkg/search: enable using Around with unsorted blob sources
The Around Query field was only taken into account if the source of blobs
to search through was sorted (e.g. a slice of sorted permanodes from the
corpus).

However, since (when requested) we already do sort all the matching
blobs when they don't come from a sorted source, and we slice the
results to a range matching the requested limit, there is nothing
preventing from centering that final slice on the requested "around".

In addition to the intrinsic feature, this offers some sort of
continuation mechanism, which makes up for the fact that using a
continuation token is also only possible in some limited cases.

This will be useful in particular for file searches, with the
DirConstraint, where we want to be able to eventually get all the
results, but in several limited batches (e.g. from the search session of
the web UI).

Change-Id: Ic9c10cac9670b43b35994e3eff13221728611c70