pkg/search: support location area predicate
This change adds the "locrect" search predicate, which works like the
"loc" predicate, except it allows to specify the coordinates of a
rectangular area as a location, instead of a named location.

The coordinates are the latitude and longitude of the North-East corner
of the rectangle, followed by the latitude and longitude of the
South-West corner. For example: locrect:48.63,-123.37,46.59,-121.28

Related: issue #934

Change-Id: I0cf39c1d0b49d557b2081f07b2c8b4508ccfc052