pkg/search: add Map sort
Since the map aspect was added to the web UI, it was discovered that it
currently does not scale well with the number of matching nodes.

The actual reason is that the search session is requesting an ever
increasing window of results, to get all results, instead of taking into
account the Continue token or Around mechanism.

However, this bug gives the opportunity to optimize the results for this
kind of interface. Instead of requesting the results in creation order,
until we get them all, we can request the set of nodes that looks the
best when displayed on the map. In other words, if there are more
results that the requested limit, the selected set of nodes should be
one that spreads around on the relevant area as evenly as possible.

This kind of selection is implemented in this CL, and will be used by
queries specifying the MapSort sort type.

Related: Issue #934

Change-Id: I6eb4901b40332863f17dab1ec4bfc11f3e99092a