pkg/search: fix Around algo with unsorted source
The search for the Around blobRef was using the equality operator,
instead of an ordering operator, which was incorrect.

Also, if the Around blob.Ref had not been found during the search, we
were only returning early for special cases, which is incorrect. As soon
as we know that an Around blob is wanted, and we know it has not been
found, the search should abort.

Relatedly, a formal check to enforce that Around is incompatible with
MapSorted was added.

Finally, some tests to check the Around algorithm, in the case where we
don't use a sorted source, were added.

Change-Id: I94fc984cf4130badc879cdadaba718ef6361c9b7