client/android: fix container detection for docker in systemd cgroup
It seems that on other (more modern?) configurations, the cgroup for
processes in docker is /system.slice/docker.service instead of /docker.

I have not been able to reproduce such a configuration. But until we have
a better way to detect whether a process is in docker, it's probably ok
to just add the new path to validate the detection, which this change
does.

Fixes #973

Change-Id: I9d52bcd79a906fdc1f8310b2c8063930cf9807fb