clients/android: for building, make caching volume dirs on the host
The ~/.gradle and ~/.android dirs are mounted as volumes for the docker
container when building the app, because we want the gradle and android
deps to be cached on the host after they've been downloaded.

The build process was assuming that these dirs always already existed,
and this assumption makes the building process fail if they actually
don't.

Therefore this change creates the directories on the host if they don't
already exist.

Change-Id: I90b5f73db0ffcbe76adc12157e142091808bde2b