camlistored/ui: reset new attribute row on navigation
When on the permanode details aspect, and navigating from one permanode
to the next, if some change had already been started on the new
attribute row, the change would "propagate" to the next permanode, which
is not desirable.

That is because the state of the new attribute row was not reset on
navigation, which this change does.

Fixes #923

Change-Id: I6dee8763b2c1fcac4b3592b21adc954eab9d0ffe