camweb: make syncing to github concurrent
Camweb is in charge of syncing the commits from the gerrit repository to
our mirror on github.

It also sends the commits as e-mails to the Camlistore mailing-list.

The former was only happening after the latter did, but there was no
reason to do it that way, as these tasks are independent. Moreover, when
sending e-mails is problematic (issue #980), it delays syncing to
github.

This change therefore makes syncing to github concurrent with
sending the e-mails.

Change-Id: I63a2b3f5b49df58ca30ec5153ce65eafb44b5b28