camweb: enable default contributor name
Because I forgot to configure git one time on another machine, one of my
commits is with a lame username, and that username gets picked (instead
of the other ones) as my identity when generating the contributors page.

Therefore, this change allows to specify a default name among the already
hardcoded information for some of the contributors, so I can override
the badly configured identity that would appear otherwise.

Change-Id: I60dfe60f83d20aa33f7bf425e5b84a34bbd5a0b1