vendor: avoid retaining irrelevant vendor files and enumerate existing removals
Previous vendoring removed a number of files manually.  This enumerates the
existing removals in .gitignore and makes it easier to transition to another
dependency management system. We also take this opportunity to add travis and
gitignore files to the list and remove those from the vendor tree.

Part of solving #889

Change-Id: Ia1b1b11ea5f2065de3a069022c2ee346e2befe78