pkg/serverinit: keep DBNames compatibility for existing GCE instances
followup of 7eda9fd5027fda88166d6c03b6490cffbf2de5fb

We want existing Perkeep instances on GCE to be able to keep on running
with their DBNames-style existing databases.

To that end, we introduce the "perkeep-config-version" metadata key,
which will be set by the launcher from now on.

When perkeep configuration starts, it can lookup that key. If it is set,
it means we're in a newly created instance, and we don't need to care
about DBNames compatibility. If not, we modify the low level
configuration on the fly, so that it keeps on using the old DBNames
values that were set for a GCE instance.

Change-Id: I611811fdb9c68777c2ba799e9047d00ec0bae040